requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-redis==5.4.0
redis==5.0.1
razorpay==1.4.1
mysqlclient==2.2.0
gunicorn==21.2.0
python-decouple==3.8
